requirements.txt
inquirer~=3.1.3
termcolor~=2.3.0
art~=6.1
loguru~=0.7.2
pydantic~=2.5.2
pydantic_settings~=2.1.0
python-binance~=1.0.19
pandas~=2.1.3